add loading spinner and refactor things
