+++ packages/shared-components/src/components/button/index.js
@@ -1,4 +1,4 @@
-import React, { useState, useRef, useEffect } from 'react';
+import React from 'react';
 import { Arrow as ArrowIcon } from '../icon';
 import Link from 'gatsby-link';
 
@@ -32,121 +32,6 @@ export const OvalSimple = ({ children, onClick }) => (
     {children}
   </button>
 );
-
-export const ProductCta = ({
-  children,
-  handleEmailSubmit,
-  productName,
-  buttonText,
-}) => {
-  const [showInput, setShowInput] = useState(false);
-  const [email, setEmail] = useState('');
-  const [buttonWidth, setButtonWidth] = useState(null);
-  const [statusMsg, setStatusMsg] = useState('');
-
-  const inputRef = useRef(null);
-  const buttonRef = useRef(null);
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    const response = await handleEmailSubmit(productName, email);
-    if (response?.status === 200) {
-      setEmail('');
-      setStatusMsg('success');
-      return;
-    }
-
-    setStatusMsg('error');
-  };
-
-  const handleRevealInput = () => {
-    setShowInput(true);
-  };
-
-  // Set initial width to facilitate for transition effect
-  useEffect(() => {
-    if (buttonRef) {
-      setButtonWidth(buttonRef.current.offsetWidth + 'px');
-    }
-  }, [productName]);
-
-  useEffect(() => {
-    if (!showInput && buttonWidth) {
-      buttonRef.current.style.width = buttonWidth;
-    }
-    if (showInput && buttonWidth) {
-      buttonRef.current.style.width = '330px';
-      inputRef.current.focus();
-    }
-  }, [buttonWidth, showInput]);
-
-  return (
-    <div className="my-6">
-      <div className="relative inline-block">
-        <button
-          data-testid="mail-reveal-btn"
-          ref={buttonRef}
-          className={`uppercase tracking-wider text-base px-8 py-6px text-navy border border-navy rounded-full font-semibold focus:outline-none ease-all transition duration-500 ease-all ${
-            showInput && 'pointer-events-none'
-          }`}
-          style={{
-            borderWidth: '2px',
-            transition: '0.5s ease all',
-            height: '42px',
-          }}
-          onClick={handleRevealInput}
-          aria-label={children}
-        >
-          <span
-            className={`transition duration-500 whitespace-nowrap valid:text-green invalid:text-red ${
-              showInput && 'opacity-0 '
-            } h-full`}
-          >
-            {buttonText.en}
-          </span>
-        </button>
-        <form onSubmit={handleSubmit}>
-          <input
-            ref={inputRef}
-            className={`absolute left-0 bg-none bottom-1/2 transform pl-8 translate-y-1/2 w-72 focus:outline-none ${
-              !showInput && 'pointer-events-none opacity-0'
-            }`}
-            style={{ background: 'none' }}
-            placeholder="Enter your email here..."
-            value={email}
-            type="email"
-            required
-            onChange={(e) => setEmail(e.target.value)}
-          />
-          <div
-            className={`absolute right-0 bottom-1/2 transform translate-y-1/2 w-18 bg-navy rounded-r-full h-full flex items-center transition duration-500 opacity-0 ${
-              showInput ? 'opacity-100' : 'pointer-events-none'
-            }`}
-          >
-            <button
-              data-testid="mail-submit-btn"
-              type="submit"
-              className="text-white ml-3 focus:outline-none"
-            >
-              SEND
-            </button>
-          </div>
-        </form>
-        <div
-          className={`${
-            statusMsg ? 'opacity-100' : 'opacity-0 pointer-events-none'
-          } ${statusMsg === 'success' && 'bg-navy'} ${
-            statusMsg === 'error' && 'bg-red-700'
-          }   absolute h-full w-full top-0 z-20 rounded-full text-white flex items-center justify-center font-bold transition duration-300`}
-        >
-          {statusMsg === 'success'
-            ? 'Successfully sent!'
-            : 'Something went wrong!'}
-        </div>
-      </div>
-    </div>
-  );
-};
 
 export const CtaArrow = ({ children, onClick, path }) => (
   <Link to={path || ''}>
